better int overflow protection

+++ rolling_mean.cpp
@@ -1,18 +1,20 @@
 #include <cstdlib>
 #include <ctime>
 #include <iostream>
+
+#include <inttypes.h>
 #include <unistd.h>
 
 int main(void) 
 {
-    long prev_mean = 0;
+    int prev_mean = 0;
     size_t n = 0;
 
     srand((unsigned int)time(NULL));
 
     for (;;) {
         int x = rand() % 100;
-        long mean = (prev_mean * n + x) / (n + 1);
+        int mean = ((intmax_t)prev_mean * n + x) / (n + 1);
         prev_mean = mean;
         ++n;
